Add function to get the bit at a specified index

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int mask = 1 << index;
+	unsigned long int mask = 1 << index;
 
 	if (n & mask)
 		return (1);
